Fixing documentation warnings. Moving MutableSpanExtensions into ints own file.
